# docs/apidog-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FundLedger API",
    "version": "0.1.0",
    "description": "Apidog-importable OpenAPI collection for the FundLedger multi-tenant SaaS API."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    },
    {
      "url": "https://api.fundledger.app",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/v1/schedules": {
      "get": {
        "tags": ["Schedules"],
        "summary": "List schedules",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      },
      "post": {
        "tags": ["Schedules"],
        "summary": "Create schedule and generate fixed per-member dues",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/ScheduleCreate" },
        "responses": { "201": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/Error" } }
      }
    },
...
    "/v1/schedules/{id}": {
      "patch": {
        "tags": ["Schedules"],
        "summary": "Edit or close schedule",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }, { "$ref": "#/components/parameters/IdPath" }],
        "requestBody": { "$ref": "#/components/requestBodies/ScheduleUpdate" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/Error" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ProvisioningApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Billing/provisioning secret configured via PROVISIONING_API_KEYS."
      },
      "ProvisioningAuthorization": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Alternative form: Authorization: ApiKey <key>."
      }
    },
    "parameters": {
      "ProjectIdHeader": {
        "name": "X-Project-Id",
        "in": "header",
        "required": false,
        "schema": { "type": "string" },
        "description": "Active project scope. If omitted, the session active project is used."
      },
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      }
    },
    "responses": {
      "Success": {
        "description": "Success response envelope",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/SuccessEnvelope" }
          }
        }
      },
      "Error": {
        "description": "Error response envelope",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      }
    },
    "requestBodies": {
...
      "ScheduleCreate": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ScheduleCreate" },
            "example": { "name": "Roofing Assessment", "unit_amount": 50000, "due_date": "2026-11-20" }
          }
        }
      },
      "ScheduleUpdate": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ScheduleUpdate" },
            "example": { "status": "closed" }
          }
        }
      },
...
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": ["ok", "data"],
        "properties": {
          "ok": { "type": "boolean", "example": true },
          "data": { "type": "object", "additionalProperties": true },
          "meta": { "type": "object", "additionalProperties": true }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": ["ok", "error"],
        "properties": {
          "ok": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string", "example": "VALIDATION" },
              "message": { "type": "string", "example": "Invalid request" },
              "fields": { "type": "object", "additionalProperties": true }
            }
          }
        }
      },
...
      "ScheduleStatus": {
        "type": "string",
        "enum": ["draft", "active", "closed"]
      },
...
      "PenaltyPolicy": {
        "type": "object",
        "required": ["enabled", "grace_months"],
        "properties": {
          "enabled": { "type": "boolean" },
          "grace_months": { "type": "integer", "minimum": 0, "maximum": 2 },
          "onetime_percent": { "type": "number", "minimum": 0, "maximum": 100 },
          "recurring_percent": { "type": "number", "minimum": 0, "maximum": 100 },
          "recurring_period": { "type": "string", "enum": ["weekly", "monthly"] },
          "compounding": { "type": "boolean" },
          "max_percent": { "type": "number", "minimum": 0, "maximum": 100 }
        }
      },
...
      "ScheduleCreate": {
        "type": "object",
        "required": ["name", "unit_amount", "due_date"],
        "properties": {
          "name": { "type": "string" },
          "unit_amount": { "type": "integer", "minimum": 1 },
          "due_date": { "type": "string", "format": "date" },
          "status": { "$ref": "#/components/schemas/ScheduleStatus" },
          "penalty_policy": { "$ref": "#/components/schemas/PenaltyPolicy" }
        }
      },
...
      "ScheduleUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "due_date": { "type": "string", "format": "date" },
          "status": { "$ref": "#/components/schemas/ScheduleStatus" },
          "penalty_policy": { "$ref": "#/components/schemas/PenaltyPolicy" }
        }
      },
...
    }
  }
}